refactor: move solver into separate svc
